--- a/asv_bench/benchmarks/stat_ops.py
+++ b/asv_bench/benchmarks/stat_ops.py
@@ -95,7 +95,7 @@ class Rank:
 
 class Correlation:
 
-    params = [["spearman"]]
+    params = [["spearman", "kendall", "pearson"]]
     param_names = ["method"]
 
     def setup(self, method):