@jbrockmendel 2021-06-07
> exposing some new kwarg to allow cython parallelism

What is the scenario in which prange is available but the user would want to disable it?