@mzeitlin11 2021-06-07
> > exposing some new kwarg to allow cython parallelism
> 
> What is the scenario in which prange is available but the user would want to disable it?

Sorry "allow" was a poor word choice.  More specifically, I think it would make sense for `prange` to default to `num_threads=1`, with the user then able to specify using more threads.

By default `prange` would defer to OpenMP to choose the number of threads, which would usually set the value to the number of cores, which can be problematic for cases like batch computing. You could probably manually set an environment variable to get around that on the user-end, but having the default behavior to use all cores would probably be a surprising change that should be opt-in.