@jreback 2021-06-08
numba will just work for parallelism

+1 on adding similar apis as we have for window functions

cc @mroeschke