@mroeschke 2021-06-08
> EDIT: also agree numba would make things easier, do you know if there has been any exploration of using numba instead of cython?

All window aggregations and groupby transform/agg have API support for a `engine` argument that can accept `'numba'` as an argument which will do the computation using numba. It essentially means rewriting the cython code back to python and adding the `jit` decorator.

You can see `pandas/core/window/numba_.py` and `pandas/core/groupby/numba_.py` for the pattern I was using to implement these.