@mzeitlin11 2021-06-09
> some followon suggestions.

Will look into removing that gil section. The main complication is that `rank_1d` requires a `np.lexsort` step that would have to be done with the gil (or perhaps some kind of presorting could be done, will look into it)